# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  nouveau: 'nouveau',
  contacte: 'contacte',
  qualifie: 'qualifie',
  proposition: 'proposition',
  gagne: 'gagne',
  perdu: 'perdu',
} as const;

export interface Lead {
  id: number;
  nom: string;
  /** @nullable */
  entreprise?: string | null;
  courriel: string;
  /** @nullable */
  telephone?: string | null;
  service: string;
  /** @nullable */
  message?: string | null;
  status: LeadStatus;
  createdAt: string;
  updatedAt: string;
}

export interface LeadInput {
  /** @minLength 1 */
  nom: string;
  entreprise?: string;
  /** @minLength 1 */
  courriel: string;
  /** @minLength 1 */
  telephone: string;
  /** @minLength 1 */
  service: string;
  message?: string;
}

export interface LeadUpdate {
  status: LeadStatus;
}

export interface LeadsSummaryGroup {
  status: LeadStatus;
  count: number;
}

export interface LeadsSummary {
  total: number;
  groups: LeadsSummaryGroup[];
}

export interface Visit {
  id: number;
  sessionId: string;
  startedAt: string;
  lastSeenAt: string;
  durationSeconds: number;
  converted: boolean;
}

export interface VisitInput {
  /** @minLength 1 */
  sessionId: string;
}

export interface VisitUpdate {
  /** @minimum 0 */
  durationSeconds?: number;
  converted?: boolean;
}

export interface VisitWindowStats {
  count: number;
  avgDurationSeconds: number;
  converted: number;
  conversionRate: number;
}

export interface VisitStats {
  last24h: VisitWindowStats;
  last7d: VisitWindowStats;
}