# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Prospect / pipeline operations
  - name: visits
    description: Website traffic tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit the public contact form as a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Invalid input
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads (admin only)
      security:
        - adminPassword: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
  /leads/summary:
    get:
      operationId: getLeadsSummary
      tags: [leads]
      summary: Pipeline summary grouped by status (admin only)
      security:
        - adminPassword: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsSummary"
        "401":
          description: Unauthorized
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead's pipeline status (admin only)
      security:
        - adminPassword: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead (admin only)
      security:
        - adminPassword: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /visits:
    post:
      operationId: createVisit
      tags: [visits]
      summary: Start tracking a new visitor session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "400":
          description: Invalid input
  /visits/{sessionId}:
    patch:
      operationId: updateVisit
      tags: [visits]
      summary: Update a visitor session's time spent and/or conversion status
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "400":
          description: Invalid input
        "404":
          description: Not found
  /visits/stats:
    get:
      operationId: getVisitStats
      tags: [visits]
      summary: Traffic analytics summary (admin only)
      security:
        - adminPassword: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitStats"
        "401":
          description: Unauthorized
components:
  securitySchemes:
    adminPassword:
      type: apiKey
      in: header
      name: X-Admin-Password
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LeadStatus:
      type: string
      enum: [nouveau, contacte, qualifie, proposition, gagne, perdu]
    Lead:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        entreprise:
          type: ["string", "null"]
        courriel:
          type: string
        telephone:
          type: ["string", "null"]
        service:
          type: string
        message:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LeadStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nom
        - courriel
        - service
        - status
        - createdAt
        - updatedAt
    LeadInput:
      type: object
      properties:
        nom:
          type: string
          minLength: 1
        entreprise:
          type: string
        courriel:
          type: string
          minLength: 1
        telephone:
          type: string
          minLength: 1
        service:
          type: string
          minLength: 1
        message:
          type: string
      required:
        - nom
        - courriel
        - telephone
        - service
    LeadUpdate:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/LeadStatus"
      required:
        - status
    LeadsSummaryGroup:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/LeadStatus"
        count:
          type: integer
      required:
        - status
        - count
    LeadsSummary:
      type: object
      properties:
        total:
          type: integer
        groups:
          type: array
          items:
            $ref: "#/components/schemas/LeadsSummaryGroup"
      required:
        - total
        - groups
    Visit:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        startedAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
        converted:
          type: boolean
      required:
        - id
        - sessionId
        - startedAt
        - lastSeenAt
        - durationSeconds
        - converted
    VisitInput:
      type: object
      properties:
        sessionId:
          type: string
          minLength: 1
      required:
        - sessionId
    VisitUpdate:
      type: object
      properties:
        durationSeconds:
          type: integer
          minimum: 0
        converted:
          type: boolean
      minProperties: 1
    VisitWindowStats:
      type: object
      properties:
        count:
          type: integer
        avgDurationSeconds:
          type: integer
        converted:
          type: integer
        conversionRate:
          type: number
      required:
        - count
        - avgDurationSeconds
        - converted
        - conversionRate
    VisitStats:
      type: object
      properties:
        last24h:
          $ref: "#/components/schemas/VisitWindowStats"
        last7d:
          $ref: "#/components/schemas/VisitWindowStats"
      required:
        - last24h
        - last7d