# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  sessionId: string;
  startedAt: Date;
  lastSeenAt: Date;
  durationSeconds: number;
  converted: boolean;
}