# lib/api-zod/src/generated/types/visitStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VisitWindowStats } from './visitWindowStats';

export interface VisitStats {
  last24h: VisitWindowStats;
  last7d: VisitWindowStats;
}