# lib/api-zod/src/generated/types/visitUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisitUpdate {
  /** @minimum 0 */
  durationSeconds?: number;
  converted?: boolean;
}